app.go refactored
